:art: :recycle: Refactoring architecture
