+++ packages/app/src/App.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useFirebaseAuth } from './firebase';
-import { LoginPage } from './LoginPage'
-import { Roulette } from './Roulette'
+import LoginPage from '../src/components/pages/LoginPage/index'
+import Roulette from '../src/components/pages/RoulettePage/index'
 
 function App() {
   const { isSignedIn, loading } = useFirebaseAuth()
-  
+
   if (loading) return 'Chargement...'
 
   return (
